remove dependency on pretty-data
The package pretty-data is no longer maintained and causes some problems
for modern build tools. Opted to not replace it with another package, since
the formatting is optional (all tests are still working) and could be handled by
the consumer of this package if required.

Related to https://github.com/bjnortier/dxf/issues/85